requirements.txt
attrs==24.1.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
numpy==2.0.1
outcome==1.3.0.post0
pandas==2.2.2
pycparser==2.22
pymongo==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.66.5
trio==0.26.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
user_agent==0.1.10
websocket-client==1.8.0
wsproto==1.2.0
